Evaluation of Individual-Level Surrogacy for Survival and Ordinal Endpoints

This function can estimate Kendall's \(\tau\) as follows: Oakes (1982)'s non-parametric estimator \(\hat{\tau}_o\), Lakhal et al. (2009)'s inverse probability weighting (IPCW) estimators \(\hat{\tau}_{mo1}\) and \(\hat{\tau}_{mo2}\), and Okui et al. (2024)'s modified IPCW estimators \(\hat{\tau}_{so1}\) and \(\hat{\tau}_{so2}\). \(\hat{\tau}_{so1}\) and \(\hat{\tau}_{so2}\) can take into account tie data. Confidence intervals were based on jackknife standard errors with a normal approximation.

Usage

surrosurvo(y, event, x, level = 0.95, confint = TRUE, parallel = FALSE)

Arguments

y

the survival time or censoring time outcome vector

event

the event indicator outcome vector

x

the ordinal (or continuous) outcome vector

level

the confidence level of the confidence interval (default = 0.95)

confint

whether the calculation of confidence intervals is performed (default = TRUE)

parallel

the number of threads used in parallel computing, or FALSE that means single threading (default = FALSE)

Value

method: names for estimators of \(\hat{\tau}\)

tau: estimates \(\hat{\tau}\)

se: standard errors for \(\hat{\tau}\)

lcl: lower confidence limits \(\hat{\tau}_l\)

ucl: uper confidence limits \(\hat{\tau}_u\)

References

Oakes, D. (1982). A model for association in bivariate survival data. J R Stat Soc Ser B Methodol. 44(3): 414-422. https://doi.org/10.1111/j.2517-6161.1982.tb01222.x

Lakhal, L, Rivest, L-P., and Beaudoin, D. (2009). IPCW estimator for Kendall's tau under bivariate censoring. Int J Biostat. 5(1): Article 8. https://doi.org/10.2202/1557-4679.1121

Okui, J., Nagashima, K., Matsuda, S., et al (2024). Surrogacy between pathological complete response and overall survival: An individual patient data analysis of ten RCTs on neoadjuvant treatment plus surgery for esophageal cancer. In preparation.

See also

Examples

library(surrosurvo)
set.seed(1234)
data <- generate_rv(0.7, 300, 0.3)
surrosurvo(data$y, data$event, data$x)
#>   method       tau         se       lcl       ucl
#> 1 taumo1 0.6705591 0.02860693 0.6144905 0.7266276
#> 2 taumo2 0.6719080 0.01936183 0.6339595 0.7098565
#> 3   tauo 0.5268004 0.02155815 0.4845472 0.5690537
#> 4 tauso1 0.7541218 0.03217435 0.6910613 0.8171824
#> 5 tauso2 0.7620473 0.03099643 0.7012954 0.8227991

# Parallel computation
surrosurvo(data$y, data$event, data$x, parallel = 2)
#>   method       tau         se       lcl       ucl
#> 1 taumo1 0.6705591 0.02860693 0.6144905 0.7266276
#> 2 taumo2 0.6719080 0.01936183 0.6339595 0.7098565
#> 3   tauo 0.5268004 0.02155815 0.4845472 0.5690537
#> 4 tauso1 0.7541218 0.03217435 0.6910613 0.8171824
#> 5 tauso2 0.7620473 0.03099643 0.7012954 0.8227991